requirements.txt
transformers
pandas
streamlit
python-dotenv
PyPDF2
langchain-community
langchain
openai
faiss-cpu
huggingface_hub
langchain_openai
sentence-transformers
langchain_huggingface
